renames package to hecdss
